requirements.txt
slack-bolt==1.19.0  # Slack Bolt framework
slack-sdk==3.31.0   # Slack SDK (included via slack-bolt, but explicit for clarity)
python-dotenv==1.0.1  # For loading .env files
pymilvus==2.6.2
openai>=1.3.0,<2.0.0  # OpenAI API for embeddings (compatible range)